move cloudagent connection helper into connection helper
