+++ test/helpers/connection.ts
@@ -352,3 +352,22 @@ export const withBobAndCharlie = function (context: PartialQuery.Context) {
     await cleanupConnections(context.agent.charlie, context.db.charlie)
   })
 }
+
+export async function withAliceReceivesBobsInvite(context: {
+  app: express.Express
+  remoteCloudagent: VeritableCloudagent
+}) {
+  const invite = await context.remoteCloudagent.createOutOfBandInvite({ companyName: alice.company_name })
+  const inviteContent = Buffer.from(
+    JSON.stringify({
+      companyNumber: alice.company_number,
+      inviteUrl: invite.invitationUrl,
+    }),
+    'utf8'
+  ).toString('base64url')
+
+  return await post(context.app, '/connection/new/receive-invitation', {
+    invite: inviteContent,
+    action: 'createConnection',
+  })
+}
