+++ test/integration/newConnection.test.ts
@@ -8,8 +8,8 @@ import Database from '../../src/models/db/index.js'
 import EmailService from '../../src/models/emailService/index.js'
 import VeritableCloudagent from '../../src/models/veritableCloudagent/index.js'
 import createHttpServer from '../../src/server.js'
-import { cleanupCloudagent, withAliceReceivesBobsInvite } from '../helpers/cloudagent.js'
-import type { TwoPartyConnection } from '../helpers/connection.js'
+import { cleanupCloudagent } from '../helpers/cloudagent.js'
+import { TwoPartyConnection, withAliceReceivesBobsInvite } from '../helpers/connection.js'
 import { cleanupDatabase } from '../helpers/db.js'
 import { alice, bobDbConfig, mockEnvBob } from '../helpers/fixtures.js'
 import { mockLogger } from '../helpers/logger.js'
